feat: update ChatModal component with improved message handling and loading state
